Cleanups per code review comments. IWYU
