+++ cpp/src/arrow/util/compression.h
@@ -22,8 +22,8 @@
 #include <memory>
 #include <string>
 
+#include "arrow/result.h"
 #include "arrow/status.h"
-#include "arrow/type_fwd.h"
 #include "arrow/util/visibility.h"
 
 namespace arrow {
@@ -173,56 +173,6 @@ class ARROW_EXPORT Codec {
   /// \brief Initializes the codec's resources.
   virtual Status Init();
 };
-
-// ----------------------------------------------------------------------
-// Codec factories
-
-namespace internal {
-
-// Brotli compression quality is max (11) by default, which is slow.
-// We use 8 as a default as it is the best trade-off for Parquet workload.
-constexpr int kBrotliDefaultCompressionLevel = 8;
-
-// Brotli codec.
-std::unique_ptr<Codec> MakeBrotliCodec(
-    int compression_level = kBrotliDefaultCompressionLevel);
-
-// BZ2 codec.
-constexpr int kBZ2DefaultCompressionLevel = 9;
-std::unique_ptr<Codec> MakeBZ2Codec(int compression_level = kBZ2DefaultCompressionLevel);
-
-// GZip
-constexpr int kGZipDefaultCompressionLevel = 9;
-
-struct GZipFormat {
-  enum type {
-    ZLIB,
-    DEFLATE,
-    GZIP,
-  };
-};
-
-std::unique_ptr<Codec> MakeGZipCodec(int compression_level = kGZipDefaultCompressionLevel,
-                                     GZipFormat::type format = GZipFormat::GZIP);
-
-// Snappy
-std::unique_ptr<Codec> MakeSnappyCodec();
-
-// Lz4 "raw" format codec.
-std::unique_ptr<Codec> MakeLz4RawCodec();
-
-// Lz4 frame format codec.
-std::unique_ptr<Codec> MakeLz4FrameCodec();
-
-// ZSTD codec.
-
-// XXX level = 1 probably doesn't compress very much
-constexpr int kZSTDDefaultCompressionLevel = 1;
-
-std::unique_ptr<Codec> MakeZSTDCodec(
-    int compression_level = kZSTDDefaultCompressionLevel);
-
-}  // namespace internal
 
 }  // namespace util
 }  // namespace arrow
