+++ cpp/src/arrow/util/compression_brotli.cc
@@ -15,10 +15,11 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "arrow/util/compression.h"
+#include "arrow/util/compression_internal.h"
 
 #include <cstddef>
 #include <cstdint>
+#include <memory>
 
 #include <brotli/decode.h>
 #include <brotli/encode.h>
@@ -32,6 +33,8 @@
 namespace arrow {
 namespace util {
 namespace internal {
+
+namespace {
 
 class BrotliDecompressor : public Decompressor {
  public:
@@ -179,12 +182,35 @@ class BrotliCodec : public Codec {
   }
 
   Result<int64_t> Decompress(int64_t input_len, const uint8_t* input,
-                             int64_t output_buffer_len, uint8_t* output_buffer) override;
+                             int64_t output_buffer_len, uint8_t* output_buffer) override {
+    DCHECK_GE(input_len, 0);
+    DCHECK_GE(output_buffer_len, 0);
+    std::size_t output_size = static_cast<size_t>(output_buffer_len);
+    if (BrotliDecoderDecompress(static_cast<size_t>(input_len), input, &output_size,
+                                output_buffer) != BROTLI_DECODER_RESULT_SUCCESS) {
+      return Status::IOError("Corrupt brotli compressed data.");
+    }
+    return output_size;
+  }
+
+  int64_t MaxCompressedLen(int64_t input_len,
+                           const uint8_t* ARROW_ARG_UNUSED(input)) override {
+    DCHECK_GE(input_len, 0);
+    return BrotliEncoderMaxCompressedSize(static_cast<size_t>(input_len));
+  }
 
   Result<int64_t> Compress(int64_t input_len, const uint8_t* input,
-                           int64_t output_buffer_len, uint8_t* output_buffer) override;
-
-  int64_t MaxCompressedLen(int64_t input_len, const uint8_t* input) override;
+                           int64_t output_buffer_len, uint8_t* output_buffer) override {
+    DCHECK_GE(input_len, 0);
+    DCHECK_GE(output_buffer_len, 0);
+    std::size_t output_size = static_cast<size_t>(output_buffer_len);
+    if (BrotliEncoderCompress(compression_level_, BROTLI_DEFAULT_WINDOW,
+                              BROTLI_DEFAULT_MODE, static_cast<size_t>(input_len), input,
+                              &output_size, output_buffer) == BROTLI_FALSE) {
+      return Status::IOError("Brotli compression failure.");
+    }
+    return output_size;
+  }
 
   Result<std::shared_ptr<Compressor>> MakeCompressor() override {
     auto ptr = std::make_shared<BrotliCompressor>(compression_level_);
@@ -204,37 +230,7 @@ class BrotliCodec : public Codec {
   int compression_level_;
 };
 
-Result<int64_t> BrotliCodec::Decompress(int64_t input_len, const uint8_t* input,
-                                        int64_t output_buffer_len,
-                                        uint8_t* output_buffer) {
-  DCHECK_GE(input_len, 0);
-  DCHECK_GE(output_buffer_len, 0);
-  std::size_t output_size = static_cast<size_t>(output_buffer_len);
-  if (BrotliDecoderDecompress(static_cast<size_t>(input_len), input, &output_size,
-                              output_buffer) != BROTLI_DECODER_RESULT_SUCCESS) {
-    return Status::IOError("Corrupt brotli compressed data.");
-  }
-  return output_size;
-}
-
-int64_t BrotliCodec::MaxCompressedLen(int64_t input_len,
-                                      const uint8_t* ARROW_ARG_UNUSED(input)) {
-  DCHECK_GE(input_len, 0);
-  return BrotliEncoderMaxCompressedSize(static_cast<size_t>(input_len));
-}
-
-Result<int64_t> BrotliCodec::Compress(int64_t input_len, const uint8_t* input,
-                                      int64_t output_buffer_len, uint8_t* output_buffer) {
-  DCHECK_GE(input_len, 0);
-  DCHECK_GE(output_buffer_len, 0);
-  std::size_t output_size = static_cast<size_t>(output_buffer_len);
-  if (BrotliEncoderCompress(compression_level_, BROTLI_DEFAULT_WINDOW,
-                            BROTLI_DEFAULT_MODE, static_cast<size_t>(input_len), input,
-                            &output_size, output_buffer) == BROTLI_FALSE) {
-    return Status::IOError("Brotli compression failure.");
-  }
-  return output_size;
-}
+}  // namespace
 
 std::unique_ptr<Codec> MakeBrotliCodec(int compression_level) {
   return std::unique_ptr<Codec>(new BrotliCodec(compression_level));
