+++ cpp/src/arrow/util/compression_snappy.cc
@@ -15,10 +15,11 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "arrow/util/compression.h"
+#include "arrow/util/compression_internal.h"
 
 #include <cstddef>
 #include <cstdint>
+#include <memory>
 
 #include <snappy.h>
 
@@ -33,6 +34,8 @@ namespace arrow {
 namespace util {
 namespace internal {
 
+namespace {
+
 // ----------------------------------------------------------------------
 // Snappy implementation
 
@@ -41,12 +44,40 @@ class SnappyCodec : public Codec {
   SnappyCodec() {}
 
   Result<int64_t> Decompress(int64_t input_len, const uint8_t* input,
-                             int64_t output_buffer_len, uint8_t* output_buffer) override;
+                             int64_t output_buffer_len, uint8_t* output_buffer) override {
+    size_t decompressed_size;
+    if (!snappy::GetUncompressedLength(reinterpret_cast<const char*>(input),
+                                       static_cast<size_t>(input_len),
+                                       &decompressed_size)) {
+      return Status::IOError("Corrupt snappy compressed data.");
+    }
+    if (output_buffer_len < static_cast<int64_t>(decompressed_size)) {
+      return Status::Invalid("Output buffer size (", output_buffer_len, ") must be ",
+                             decompressed_size, " or larger.");
+    }
+    if (!snappy::RawUncompress(reinterpret_cast<const char*>(input),
+                               static_cast<size_t>(input_len),
+                               reinterpret_cast<char*>(output_buffer))) {
+      return Status::IOError("Corrupt snappy compressed data.");
+    }
+    return static_cast<int64_t>(decompressed_size);
+  }
+
+  int64_t MaxCompressedLen(int64_t input_len,
+                           const uint8_t* ARROW_ARG_UNUSED(input)) override {
+    DCHECK_GE(input_len, 0);
+    return snappy::MaxCompressedLength(static_cast<size_t>(input_len));
+  }
 
   Result<int64_t> Compress(int64_t input_len, const uint8_t* input,
-                           int64_t output_buffer_len, uint8_t* output_buffer) override;
-
-  int64_t MaxCompressedLen(int64_t input_len, const uint8_t* input) override;
+                           int64_t ARROW_ARG_UNUSED(output_buffer_len),
+                           uint8_t* output_buffer) override {
+    size_t output_size;
+    snappy::RawCompress(reinterpret_cast<const char*>(input),
+                        static_cast<size_t>(input_len),
+                        reinterpret_cast<char*>(output_buffer), &output_size);
+    return static_cast<int64_t>(output_size);
+  }
 
   Result<std::shared_ptr<Compressor>> MakeCompressor() override {
     return Status::NotImplemented("Streaming compression unsupported with Snappy");
@@ -59,42 +90,7 @@ class SnappyCodec : public Codec {
   const char* name() const override { return "snappy"; }
 };
 
-Result<int64_t> SnappyCodec::Decompress(int64_t input_len, const uint8_t* input,
-                                        int64_t output_buffer_len,
-                                        uint8_t* output_buffer) {
-  size_t decompressed_size;
-  if (!snappy::GetUncompressedLength(reinterpret_cast<const char*>(input),
-                                     static_cast<size_t>(input_len),
-                                     &decompressed_size)) {
-    return Status::IOError("Corrupt snappy compressed data.");
-  }
-  if (output_buffer_len < static_cast<int64_t>(decompressed_size)) {
-    return Status::Invalid("Output buffer size (", output_buffer_len, ") must be ",
-                           decompressed_size, " or larger.");
-  }
-  if (!snappy::RawUncompress(reinterpret_cast<const char*>(input),
-                             static_cast<size_t>(input_len),
-                             reinterpret_cast<char*>(output_buffer))) {
-    return Status::IOError("Corrupt snappy compressed data.");
-  }
-  return static_cast<int64_t>(decompressed_size);
-}
-
-int64_t SnappyCodec::MaxCompressedLen(int64_t input_len,
-                                      const uint8_t* ARROW_ARG_UNUSED(input)) {
-  DCHECK_GE(input_len, 0);
-  return snappy::MaxCompressedLength(static_cast<size_t>(input_len));
-}
-
-Result<int64_t> SnappyCodec::Compress(int64_t input_len, const uint8_t* input,
-                                      int64_t ARROW_ARG_UNUSED(output_buffer_len),
-                                      uint8_t* output_buffer) {
-  size_t output_size;
-  snappy::RawCompress(reinterpret_cast<const char*>(input),
-                      static_cast<size_t>(input_len),
-                      reinterpret_cast<char*>(output_buffer), &output_size);
-  return static_cast<int64_t>(output_size);
-}
+}  // namespace
 
 std::unique_ptr<Codec> MakeSnappyCodec() {
   return std::unique_ptr<Codec>(new SnappyCodec());
